Moving some classes to styled-copmonents and changing the logo image in night-mode
