+++ src/components/Word.tsx
@@ -13,10 +13,11 @@ import {
   PlayCurrentWord,
 } from '../redux/actions/audioplayer';
 
-const WordTag = styled.a<{ highlight?: boolean }>`
+const WordTag = styled.a<{ highlight?: boolean, isNightMode?: boolean }>`
   -webkit-font-smoothing: antialiased;
-  color: ${({ highlight, theme }) =>
-    highlight ? theme.brandPrimary : theme.colors.textColor};
+  color: ${({ highlight, theme, isNightMode }) => 
+    highlight ? theme.brandPrimary : isNightMode ?
+    theme.colors.white + " !important" : theme.colors.textColor};
 `;
 
 const propTypes = {
@@ -30,6 +31,7 @@ const propTypes = {
   isCurrentVersePlaying: PropTypes.bool.isRequired,
   isSearched: PropTypes.bool,
   useTextFont: PropTypes.bool, // tmp change to compare text and code based rendering
+  isNightMode: PropTypes.bool
 };
 
 type DefaultProps = {
@@ -57,6 +59,7 @@ type Props = {
   isCurrentVersePlaying: boolean;
   isSearched?: boolean;
   useTextFont?: boolean;
+  isNightMode: boolean;
 };
 
 class Word extends Component<Props> {
@@ -119,6 +122,7 @@ class Word extends Component<Props> {
       isCurrentVersePlaying,
       word,
       useTextFont,
+      isNightMode
     } = this.props;
 
     let text = '';
@@ -145,6 +149,7 @@ class Word extends Component<Props> {
             role="button"
             tabIndex={audioPosition}
             highlight={isCurrentVersePlaying}
+            isNightMode={isNightMode}
             id={id}
             onDoubleClick={this.handleSegmentPlay}
             onClick={this.handleWordPlay}
