+++ src/styles/nightmode.scss
@@ -2,18 +2,11 @@
   background: #1d1f21 !important;
   color: #fff;
 
-  a.word,
-  a.end   {
-    color: #fff;
-  }
   nav {
     background: #32312C;
     ul li a {
       color: #fff;
     }
-  }
-  .text-translation small {
-    color: #fff;
   }
   .ayah-name-svg path {
     fill: #fff;
